added a 10-second timeout to all HTTP(S) requests
by default, requests does not provide a timeout for any calls. this would mean that in the event of an internet dropout, the library (and any application using it) would simply hang indefinitely. this resolves it by adding a 10-second timeout after which an exception will be fired by requests.